[dtensor] group all dynamo tests together
[ghstack-poisoned]
