+++ test/distributed/_tensor/test_dtensor_compile.py
@@ -5,8 +5,9 @@ import copy
 
 import torch
 import torch._dynamo
+import torch.distributed as dist
 import torch.nn as nn
-from torch.distributed._tensor import DeviceMesh
+from torch.distributed._tensor import DeviceMesh, DTensor, Replicate, Shard
 from torch.distributed.fsdp import FullyShardedDataParallel as FSDP
 from torch.distributed.tensor.parallel import PairwiseParallel, parallelize_module
 from torch.distributed.tensor.parallel.fsdp import enable_2d_with_fsdp
@@ -16,6 +17,7 @@ from torch.testing._internal.distributed._tensor.common_dtensor import (
     MLPModule,
     with_comms,
 )
+from torch.testing._internal.distributed.fake_pg import FakeStore
 
 
 class SimpleModel(nn.Module):
@@ -28,10 +30,111 @@ class SimpleModel(nn.Module):
         return self.mlp_1(self.mlp_0(input))
 
 
-class TestDTensorCompile(DTensorTestBase):
+class TestDTensorCompile(torch._dynamo.test_case.TestCase):
+    def setUp(self):
+        super().setUp()
+        fake_store = FakeStore()
+        dist.init_process_group(
+            "fake", store=fake_store, rank=0, world_size=self.world_size
+        )
+
+    def tearDown(self):
+        super().tearDown()
+        dist.destroy_process_group()
+
+    @property
+    def device_type(self) -> str:
+        return "cuda" if torch.cuda.is_available() else "cpu"
+
+    @property
+    def world_size(self) -> int:
+        return 2
+
+    def test_fakify_dtensor(self):
+        mesh = DeviceMesh(self.device_type, torch.arange(self.world_size))
+
+        # pass in DTensor as inputs/outputs to the function
+        def fn(x):
+            return x
+
+        x = DTensor.from_local(torch.rand(1), mesh, [Shard(0)], run_check=False)
+        ref = fn(x)
+
+        opt_fn = torch.compile(fn, backend="eager", fullgraph=True)
+        res = opt_fn(x)
+        self.assertEqual(res, ref)
+
+    def test_dynamo_dtensor(self):
+        mesh = DeviceMesh(self.device_type, torch.arange(self.world_size))
+
+        # test passing in DTensor as inputs/outputs and run some tensor computation
+        def fn(x):
+            return x * x + 2
+
+        x = DTensor.from_local(torch.rand(1), mesh, [Shard(0)], run_check=False)
+        ref = fn(x)
+
+        opt_fn = torch.compile(fn, backend="eager", fullgraph=True)
+        res = opt_fn(x)
+        self.assertEqual(res, ref)
+
+    def test_dynamo_dtensor_from_local(self):
+        mesh = DeviceMesh(self.device_type, torch.arange(self.world_size))
+
+        # create DTensor inside fn and run some compute
+        def fn(x):
+            dt = DTensor.from_local(x, mesh, [Replicate()], run_check=False)
+            return dt.to_local() + 2
+
+        # below is the op approach for reference
+        # from torch.distributed._tensor.api import _FromTorchTensor
+        # def from_local_tensor(x):
+        #     return _FromTorchTensor.apply(x, mesh, [Replicate()], False)
+
+        # _dt_lib_def = torch.library.Library("dtensor", "DEF")
+        # _dt_lib_def.define("from_local(Tensor self) -> Tensor")
+
+        # _dt_lib_impl = torch.library.Library("dtensor", "IMPL")
+        # _dt_lib_impl.impl("from_local", from_local_tensor, "Autograd")
+
+        x = torch.ones(1)
+        ref = fn(x)
+        opt_fn = torch.compile(fn, backend="eager", fullgraph=True)
+        res = opt_fn(x)
+        self.assertEqual(res, ref)
+
+    def test_dynamo_dtensor_from_local_redistribute(self):
+        mesh = DeviceMesh(self.device_type, torch.arange(self.world_size))
+
+        # pass in tensor as inputs/outputs, create DTensor and run redistribute
+        # (allgather collective) inside the fn
+        def fn(x):
+            dt = DTensor.from_local(x, mesh, [Shard(0)], run_check=False)
+            return dt.redistribute(mesh, [Replicate()]).to_local() + 2
+
+        x = torch.ones(1)
+        ref = fn(x)
+        opt_fn = torch.compile(fn, backend="eager", fullgraph=True)
+        res = opt_fn(x)
+        self.assertEqual(res, ref)
+
+
+class TestDTensorCompileE2E(DTensorTestBase):
     @property
     def world_size(self):
         return 4
+
+    @with_comms
+    def test_tp_compile_fullgraph(self):
+        mesh = DeviceMesh(self.device_type, torch.arange(self.world_size))
+
+        model = MLPModule(self.device_type)
+        model = parallelize_module(model, mesh, PairwiseParallel())
+        inp = torch.rand(20, 10, device=self.device_type)
+        out = model(inp)
+        compiled_mod = torch.compile(model, backend="eager", fullgraph=True)
+        compiled_out = compiled_mod(inp)
+        self.assertEqual(compiled_out, out)
 
     @with_comms
     def test_2d_fsdp_tp_compile(self):
